requirements.txt
Django>=4.2,<5.0
 djangorestframework>=3.14
 djangorestframework-simplejwt>=5.3
 django-cors-headers>=4.3
 mysqlclient>=2.2
 Pillow>=10.2
 python-dotenv>=1.0
 openai>=1.35
 requests>=2.31
 drf-spectacular>=0.27
